+++ CHANGELOG.md
@@ -7,6 +7,8 @@ v0.1.0 (in development)
   group named "`version`," the entire text matched by the regex will be used as
   the version
 - "basic" `write` method: `encoding` now defaults to UTF-8
+- New `next-version` methods: `"minor-release"`, `"smallest-release"`, and
+  `"null`"
 
 v0.1.0a1 (2021-07-05)
 ---------------------